use numpy gen

+++ declafe/feature_gen/binary/talib/AROONFeature.py
@@ -12,7 +12,7 @@ class AROONDownFeature(BinaryFeature):
     super().__init__(high, low)
 
   def bigen(self, left: pd.Series, right: pd.Series) -> pd.Series:
-    return talib.AROON(left, right, self.period)[0]
+    return talib.AROON(left.astype(float), right.astype(float), self.period)[0]
 
   def _feature_name(self) -> str:
     return f"AROONDown_{self.period}_{self.left}_{self.right}"
@@ -25,7 +25,7 @@ class AROONUpFeature(BinaryFeature):
     super().__init__(high, low)
 
   def bigen(self, left: pd.Series, right: pd.Series) -> pd.Series:
-    return talib.AROON(left, right, self.period)[1]
+    return talib.AROON(left.astype(float), right.astype(float), self.period)[1]
 
   def _feature_name(self) -> str:
     return f"AROONUp_{self.period}_{self.left}_{self.right}"
