+++ declafe/feature_gen/binary/talib/MinusDMFeature.py
@@ -15,7 +15,7 @@ class MinusDMFeature(BinaryFeature):
     self.period = period
 
   def bigen(self, left: pd.Series, right: pd.Series) -> pd.Series:
-    return talib.MINUS_DM(left, right, self.period)
+    return talib.MINUS_DM(left.astype(float), right.astype(float), self.period)
 
   def _feature_name(self) -> str:
     return f"MINUS_DM_{self.left}_{self.right}_{self.period}"
