+++ declafe/feature_gen/binary/talib/PlusDMFeature.py
@@ -15,7 +15,7 @@ class PlusDMFeature(BinaryFeature):
     self.period = period
 
   def bigen(self, left: pd.Series, right: pd.Series) -> pd.Series:
-    return talib.PLUS_DM(left, right, self.period)
+    return talib.PLUS_DM(left.astype(float), right.astype(float), self.period)
 
   def _feature_name(self) -> str:
     return f"PLUS_DM_{self.left}_{self.right}_{self.period}"
