+++ declafe/feature_gen/binary/talib/SAREXTFeature.py
@@ -7,7 +7,7 @@ from ..BinaryFeature import BinaryFeature
 class SAREXTFeature(BinaryFeature):
 
   def bigen(self, left: pd.Series, right: pd.Series) -> pd.Series:
-    return talib.SAREXT(left, right)
+    return talib.SAREXT(left.astype(float), right.astype(float))
 
   def _feature_name(self) -> str:
     return f"SAREXT_{self.left}_{self.right}"
