+++ declafe/feature_gen/binary/talib/SARFeature.py
@@ -7,7 +7,7 @@ from ..BinaryFeature import BinaryFeature
 class SARFeature(BinaryFeature):
 
   def bigen(self, left: pd.Series, right: pd.Series) -> pd.Series:
-    return talib.SAR(left, right)
+    return talib.SAR(left.astype(float), right.astype(float))
 
   def _feature_name(self) -> str:
     return f"SAR_{self.left}_{self.right}"
