+++ declafe/feature_gen/unary/talib/RSIFeature.py
@@ -16,4 +16,4 @@ class RSIFeature(UnaryFeature):
     return f"RSI_{self.period}"
 
   def gen_unary(self, ser: pd.Series) -> pd.Series:
-    return talib.RSI(ser, timeperiod=self.period)
+    return talib.RSI(ser.astype(float), timeperiod=self.period)
