+++ declafe/feature_gen/unary/talib/STOCHRSI.py
@@ -24,7 +24,7 @@ class STOCHRSIFastkFeature(UnaryFeature):
     return f"STOCHRSI_fastk_{self.period}_{self.fastk_period}_{self.fastd_period}_{self.fastd_matype}"
 
   def gen_unary(self, ser: pd.Series) -> pd.Series:
-    return talib.STOCHRSI(ser,
+    return talib.STOCHRSI(ser.astype(float),
                           timeperiod=self.period,
                           fastk_period=self.fastk_period,
                           fastd_period=self.fastd_period,
@@ -50,7 +50,7 @@ class STOCHRSIFastdFeature(UnaryFeature):
     return f"STOCHRSI_fastd_{self.period}_{self.fastk_period}_{self.fastd_period}_{self.fastd_matype}"
 
   def gen_unary(self, ser: pd.Series) -> pd.Series:
-    return talib.STOCHRSI(ser,
+    return talib.STOCHRSI(ser.astype(float),
                           timeperiod=self.period,
                           fastk_period=self.fastk_period,
                           fastd_period=self.fastd_period,
